# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChargeCreate": {
        "properties": {
          "base": {
            "default": "works_subtotal",
            "enum": [
              "works_subtotal",
              "running_total",
              "sections"
            ],
            "title": "Base",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "fixed_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixed Amount"
          },
          "kind": {
            "default": "other",
            "enum": [
              "contingency",
              "work_charged_establishment",
              "labour_cess",
              "seigniorage",
              "royalty",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "percentage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentage"
          },
          "section_keys": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 200,
            "title": "Section Keys",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "ChargeCreate",
        "type": "object"
      },
...
      "ChargePatch": {
        "properties": {
          "base": {
            "anyOf": [
              {
                "enum": [
                  "works_subtotal",
                  "running_total",
                  "sections"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "fixed_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixed Amount"
          },
          "kind": {
            "anyOf": [
              {
                "enum": [
                  "contingency",
                  "work_charged_establishment",
                  "labour_cess",
                  "seigniorage",
                  "royalty",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "percentage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentage"
          },
          "section_keys": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 200,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Keys"
          }
        },
        "title": "ChargePatch",
        "type": "object"
      },
...
      "DefaultsIn": {
        "properties": {
          "charges": {
            "items": {
              "$ref": "#/components/schemas/DefaultChargeIn"
            },
            "maxItems": 20,
            "title": "Charges",
            "type": "array"
          },
          "gst": {
            "$ref": "#/components/schemas/GstIn"
          },
          "rounding": {
            "default": "nearest_rupee",
            "enum": [
              "none",
              "nearest_rupee",
              "nearest_10",
              "nearest_100",
              "nearest_1000"
            ],
            "title": "Rounding",
            "type": "string"
          }
        },
        "required": [
          "gst"
        ],
        "title": "DefaultsIn",
        "type": "object"
      },
...
      "Envelope_AbstractOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AbstractOut"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[AbstractOut]",
        "type": "object"
      },
...
      "Envelope_DefaultsOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DefaultsOut"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[DefaultsOut]",
        "type": "object"
      },
...
      "Envelope_ValidationOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ValidationOut"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ValidationOut]",
        "type": "object"
      },
...
      "GstIn": {
        "properties": {
          "applicable": {
            "title": "Applicable",
            "type": "boolean"
          },
          "base": {
            "default": "after_charges",
            "enum": [
              "after_charges",
              "works_subtotal"
            ],
            "title": "Base",
            "type": "string"
          },
          "mode": {
            "default": "exclusive",
            "enum": [
              "exclusive",
              "inclusive"
            ],
            "title": "Mode",
            "type": "string"
          },
          "rate_pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Pct"
          },
          "supply": {
            "default": "intra",
            "enum": [
              "intra",
              "inter"
            ],
            "title": "Supply",
            "type": "string"
          }
        },
        "required": [
          "applicable"
        ],
        "title": "GstIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "title": "Meta",
        "type": "object"
      },
...
      "RoundingIn": {
        "properties": {
          "grand_total": {
            "enum": [
              "none",
              "nearest_rupee",
              "nearest_10",
              "nearest_100",
              "nearest_1000"
            ],
            "title": "Grand Total",
            "type": "string"
          }
        },
        "required": [
          "grand_total"
        ],
        "title": "RoundingIn",
        "type": "object"
      },
...
      "app__api__v1__abstract__OrderIn": {
        "properties": {
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "OrderIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EstimateAI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/charges/{charge_id}": {
      "delete": {
        "operationId": "delete_charge_api_v1_charges__charge_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "charge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Charge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbstractOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Charge",
        "tags": [
          "abstract"
        ]
      },
      "patch": {
        "operationId": "patch_charge_api_v1_charges__charge_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "charge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Charge Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbstractOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Charge",
        "tags": [
          "abstract"
        ]
      }
    },
...
    "/api/v1/organizations/{org_id}/settings/estimate-defaults": {
      "get": {
        "operationId": "get_defaults_api_v1_organizations__org_id__settings_estimate_defaults_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DefaultsOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Defaults",
        "tags": [
          "abstract"
        ]
      },
      "put": {
        "operationId": "put_defaults_api_v1_organizations__org_id__settings_estimate_defaults_put",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DefaultsOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Defaults",
        "tags": [
          "abstract"
        ]
      }
    },
...
    "/api/v1/versions/{version_id}/abstract": {
      "get": {
        "operationId": "get_abstract_api_v1_versions__version_id__abstract_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbstractOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Abstract",
        "tags": [
          "abstract"
        ]
      }
    },
...
    "/api/v1/versions/{version_id}/charges": {
      "post": {
        "operationId": "add_charge_api_v1_versions__version_id__charges_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbstractOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Charge",
        "tags": [
          "abstract"
        ]
      }
    },
    "/api/v1/versions/{version_id}/charges/reorder": {
      "post": {
        "operationId": "reorder_charges_api_v1_versions__version_id__charges_reorder_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__abstract__OrderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbstractOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reorder Charges",
        "tags": [
          "abstract"
        ]
      }
    },
...
    "/api/v1/versions/{version_id}/gst": {
      "put": {
        "operationId": "put_gst_api_v1_versions__version_id__gst_put",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GstIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbstractOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Gst",
        "tags": [
          "abstract"
        ]
      }
    },
...
    "/api/v1/versions/{version_id}/rounding": {
      "put": {
        "operationId": "put_rounding_api_v1_versions__version_id__rounding_put",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoundingIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbstractOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Rounding",
        "tags": [
          "abstract"
        ]
      }
    },
...
    "/api/v1/versions/{version_id}/validation": {
      "get": {
        "operationId": "validate_version_api_v1_versions__version_id__validation_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ValidationOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Version",
        "tags": [
          "abstract"
        ]
      }
    },
...
  }
}